fixing useSession & improving string-avatar & rename dropDown to UserDropdown
