+++ client/src/containers/string-avatar/component.tsx
@@ -2,7 +2,10 @@ import cx from 'classnames';
 import { stringAvatarProps } from './type';
 
 const stringAvatarFn = (name: string) => {
-  const stringAvatar = `${name.split(' ')[0][0]}${name.split(' ')[1][0]}`;
+  const userNames = name.split(' ');
+  console.log(userNames);
+
+  const stringAvatar = `${userNames[0][0]}${userNames[1][0]}`;
   return stringAvatar;
 };
 
@@ -12,7 +15,7 @@ export const StringAvatar: React.FC<stringAvatarProps> = ({
 }: stringAvatarProps) => (
   <div
     className={cx(
-      `flex h-10 w-10 rounded-full items-center bg-green-600 uppercase text-white justify-center font-semibold`,
+      `flex h-10 w-10 rounded-full text-lg items-center bg-green-600 uppercase text-white justify-center font-semibold`,
       className,
     )}
   >
